validation overhaul
