+++ src/mods/ilbattle.h
@@ -6,8 +6,8 @@ namespace ilbattle {
 
 static constexpr u32 NUM_LENGTHS = 4;
 
+void track_valid_attempt();
 void tick();
 void disp();
-void new_battle();
 
 }  // namespace ilbattle
