+++ src/mods/ilmark.cpp
@@ -15,33 +15,7 @@
 namespace ilmark {
 
 static bool s_valid_run = false;
-static s16 s_paused_frame = 0;
 static bool s_is_romhack = false;
-
-static constexpr pref::BoolPref INVALID_BOOL_PREFS[] = {
-    pref::BoolPref::DisableFalloutVolumes,
-    pref::BoolPref::JumpMod,
-    pref::BoolPref::Marathon,
-    pref::BoolPref::DebugMode,
-};
-
-static constexpr pref::U8Pref INVALID_U8_PREFS[] = {
-    pref::U8Pref::TimerType,        pref::U8Pref::Friction,         pref::U8Pref::Restitution,
-    pref::U8Pref::FalloutPlaneType, pref::U8Pref::StageEditVariant,
-};
-
-void disable_invalidating_settings() {
-    // set all bool prefs to default
-    for (u8 i = 0; i < LEN(INVALID_BOOL_PREFS); i++) {
-        pref::set(INVALID_BOOL_PREFS[i], pref::get_default(INVALID_BOOL_PREFS[i]));
-    }
-    // set all u8 prefs to default
-    for (u8 i = 0; i < LEN(INVALID_U8_PREFS); i++) {
-        pref::set(INVALID_U8_PREFS[i], pref::get_default(INVALID_U8_PREFS[i]));
-    }
-
-    pref::save();
-}
 
 void init() {
     char gamecode[7] = {};
@@ -49,49 +23,16 @@ void init() {
     s_is_romhack = mkb::strcmp(gamecode, "GM2E8P") != 0;
 }
 
+void set_valid() { s_valid_run = true; }
+
 void tick() {
-    if (mkb::sub_mode == mkb::SMD_GAME_PLAY_INIT) {
-        s_valid_run = true;
-        s_paused_frame = 0;
-
-    } else if (mkb::sub_mode == mkb::SMD_GAME_PLAY_MAIN) {
-        bool paused_now = *reinterpret_cast<u32*>(0x805BC474) & 8;
-        if (paused_now) {
-            if (s_paused_frame == 0) {
-                s_paused_frame = mkb::mode_info.stage_time_frames_remaining;
-            } else {
-                s_valid_run = false;
-            }
-        }
-        // Loading savestates is disallowed
-        if (libsavest::state_loaded_this_frame()) s_valid_run = false;
-
-        // Using dpad controls is disallowed
-        bool dpad_down =
-            pad::button_down(mkb::PAD_BUTTON_DOWN) || pad::button_down(mkb::PAD_BUTTON_LEFT) ||
-            pad::button_down(mkb::PAD_BUTTON_RIGHT) || pad::button_down(mkb::PAD_BUTTON_UP);
-        if (pref::get(pref::BoolPref::DpadControls) && dpad_down) s_valid_run = false;
-
-        // Opening the mod menu is disallowed
-        if (menu_impl::is_visible()) s_valid_run = false;
-
-        // Physics must be default or custom (with all default values)
-        if (physics::using_custom_physics()) s_valid_run = false;
-
-        // Invalid bool prefs are enabled
-        for (u8 i = 0; i < LEN(INVALID_BOOL_PREFS); i++) {
-            if (pref::get(INVALID_BOOL_PREFS[i]) != pref::get_default(INVALID_BOOL_PREFS[i])) {
-                s_valid_run = false;
-            }
-        }
-        // Invalid u8 prefs are enabled
-        for (u8 i = 0; i < LEN(INVALID_U8_PREFS); i++) {
-            if (pref::get(INVALID_U8_PREFS[i]) != pref::get_default(INVALID_U8_PREFS[i])) {
-                s_valid_run = false;
-            }
-        }
-    } else if (mkb::sub_mode == mkb::SMD_GAME_GOAL_INIT) {
-        s_valid_run = s_valid_run && s_paused_frame <= mkb::mode_info.stage_time_frames_remaining;
+    if (mkb::sub_mode != mkb::SMD_GAME_PLAY_MAIN && mkb::sub_mode != mkb::SMD_GAME_GOAL_INIT &&
+        mkb::sub_mode != mkb::SMD_GAME_GOAL_MAIN &&
+        mkb::sub_mode != mkb::SMD_GAME_GOAL_REPLAY_INIT &&
+        mkb::sub_mode != mkb::SMD_GAME_GOAL_REPLAY_MAIN &&
+        mkb::sub_mode != mkb::SMD_GAME_BONUS_CLEAR_INIT &&
+        mkb::sub_mode != mkb::SMD_GAME_BONUS_CLEAR_MAIN) {
+        s_valid_run = false;
     }
 }
